WIP ckpt for PyTorch
